// src/DockiUp.Frontend/src/app/api/model/cleanupFrequency.ts
/**
 * DockiUp.API | v1
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type CleanupFrequency = 'Off' | 'Daily' | 'Weekly';

export const CleanupFrequency = {

    Off: 'Off' as CleanupFrequency,

    Daily: 'Daily' as CleanupFrequency,

    Weekly: 'Weekly' as CleanupFrequency
};