// src/DockiUp.Frontend/src/app/api/model/cleanupScheduleDto.ts
/**
 * DockiUp.API | v1
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { CleanupFrequency } from './cleanupFrequency';


export interface CleanupScheduleDto { 
    nodeId: string | null;
    frequency: CleanupFrequency;
    hourUtc: number;
    dayOfWeekUtc: number;
    pruneVolumes: boolean;
    lastRunAt: string | null;
    lastRunResult: string | null;
}
export namespace CleanupScheduleDto {
}