// src/DockiUp.Frontend/src/app/api/model/dockerImageDto.ts
/**
 * DockiUp.API | v1
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface DockerImageDto { 
    id: string;
    tags: Array<string>;
    size: number;
    created: string;
    containers: number;
    inUse: boolean;
    dangling: boolean;
}