// src/DockiUp.Frontend/src/app/api/model/dockerNetworkDto.ts
/**
 * DockiUp.API | v1
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface DockerNetworkDto { 
    id: string;
    name: string;
    driver: string;
    scope: string;
    containers: number;
    inUse: boolean;
    builtin: boolean;
}