// src/DockiUp.Frontend/src/app/api/model/dockerResourcesDto.ts
/**
 * DockiUp.API | v1
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { DiskUsageDto } from './diskUsageDto';
import { DockerImageDto } from './dockerImageDto';
import { DockerNetworkDto } from './dockerNetworkDto';
import { DockerVolumeDto } from './dockerVolumeDto';


export interface DockerResourcesDto { 
    images: Array<DockerImageDto>;
    volumes: Array<DockerVolumeDto>;
    networks: Array<DockerNetworkDto>;
    diskUsage: DiskUsageDto;
}