// src/DockiUp.Frontend/src/app/api/model/dockerVolumeDto.ts
/**
 * DockiUp.API | v1
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface DockerVolumeDto { 
    name: string;
    driver: string;
    size: number | null;
    created: string | null;
    containers: number;
    inUse: boolean;
}